fix: Remove stupid keyword.
"make the web free again" is stupid and long. This keyword does align
with my goals of making the internet a freeer place (as in freedom), but
it is just _no_.

Internet used to be free back when it started: personal websites, cool
people, it's like "by the people for the people" and now it's all
corporate bs...

Either way, removing.

Signed-off-by: Arija A. <ari@ari.lt>